third_party/blink/web_tests/platform/generic/fast/forms/select/optgroup-clicking-expected.txt

Test clicking with optgroup element

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Click enabled option
PASS $("listbox").selectedIndex is 1

Click on optgroup, should not deselect selectedIndex
PASS $("listbox").selectedIndex is 1

Click disabled option - doesn't change selectedIndex
PASS $("listbox").selectedIndex is 1

Menu list with arrow key. The test is expected to fail on OSX because of a key binding difference.
PASS $("menulist").selectedIndex is 8
PASS successfullyParsed is true

TEST COMPLETE

One
Two
Three
Four
One
Two
Three
Four


One
Two
Three
Four
One
Two
Three
Four
One
Two
Three
Four
